pyproject.toml
[project]
name = "foundry-hostedagentframework-demo"
version = "0.1.0"
description = "Seattle Hotel Agent demo using Microsoft Agent Framework and Azure AI Foundry"
requires-python = ">=3.12"
dependencies = [
    "azure-identity",
    "python-dotenv",
    "azure-ai-agentserver-agentframework==1.0.0b17",
    "azure-ai-projects>=2.0.0b4",
    "azure-search-documents>=11.6.0",
    "agent-framework-azure-ai-search>=1.0.0b260304",
]

[tool.uv]
prerelease = "allow"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
